% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compensator.R
\name{compensator}
\alias{compensator}
\alias{compensator.default}
\alias{compensator.mmpp}
\alias{compensator.hp}
\alias{compensator.mmhp}
\title{Compensators for MMHP, HP and MMPP models}
\usage{
compensator(object, t, pzt)

\method{compensator}{default}(object, t, pzt)

\method{compensator}{mmpp}(object, t, pzt)

\method{compensator}{hp}(object, t, pzt)

\method{compensator}{mmhp}(object, t, pzt)
}
\arguments{
\item{object}{a social network model}

\item{t}{event times}

\item{pzt}{probability}
}
\value{
vector of compensator
}
\description{
Compensators for MMHP, HP and MMPP models
}
