% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmhp.R
\name{mmhp}
\alias{mmhp}
\title{Create a Markov-modulated Hawkes Process(MMHP) model}
\usage{
mmhp(state, state_time, tau, lambda0, lambda1, alpha, beta)
}
\arguments{
\item{state}{different states with state = 1 corresponds to active state, the Markov process and state = 2 corresponds to inactive state, the homogeneous Poisson process}

\item{state_time}{time of each transition of Markov process}

\item{tau}{times of Poisson events}

\item{lambda0}{parameters for homogeneous Poisson process}

\item{lambda1}{parameters for Hawkes process}

\item{alpha}{parameters for Hawkes process}

\item{beta}{parameters for Hawkes process}
}
\value{
mmhp object
}
\description{
Create a Markov-modulated Hawkes Process(MMHP) model according to the given parameters: lambda0, lambda1, alpha, beta, event times and the states
}
\examples{
mmhp(c(1,2),c(0,8.36),c(0,1.27),lambda0 = 0.9, lambda1 = 1.1, alpha = 0.8, beta = 1.2)
}
