% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmhpNegLokLik.R
\name{mmhpNegLogLik}
\alias{mmhpNegLogLik}
\title{Compute negative log likelihood for Markov-modulated Hawkes Process(MMHP) model}
\usage{
mmhpNegLogLik(mmhp, t = NULL)
}
\arguments{
\item{mmhp}{a mmhp object including its Q, delta, tau, lambda0, lambda1, beta and alpha.}

\item{t}{vector containing the event times. Note that if `mmhp` obejct does have argument `tau`, then `t` is not needed anymore.}
}
\value{
a scalar indicating the negative log likelihood
}
\description{
Compute negative log likelihood for MMHP model with model specified time events or simulated time events
}
\examples{
Q <- matrix(c(-0.4, 0.4, 0.2, -0.2), ncol = 2, byrow = TRUE)
x <- mmhp(Q, delta = c(1 / 3, 2 / 3), lambda0 = 0.9, lambda1 = 1.1, alpha = 0.8, beta = 1.2)
y <- simulatemmhp(x, nsim = 5)
mmhpNegLogLik(x, y$tau)
}
