% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescaled.R
\name{rescaled}
\alias{rescaled}
\title{Generate the rescaled-inter-event-times for point process models}
\usage{
rescaled(lambda, tau)
}
\arguments{
\item{lambda}{intensity function of point process models}

\item{tau}{event times}
}
\description{
Generate the rescaled-inter-event-times for point process models in order to implement time rescaling theorem and for model checking
}
