% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snadv-package.R, R/snadv.R
\docType{package}
\name{snadv}
\alias{snadv}
\alias{snadv-package}
\title{snadv: Perform diagnosis and visualizations for social network analysis}
\description{
The model checking process is an important part in the social network analysis.
This package includes multiple methods to plot the model as well as the true data, to compare the model fit and model performace.

sandv provides a flexible grammar of social network model diagnosis. It focus on visualization tools,
intensity calculation, inspection of log likelihhood and pearson residuals.
}
\section{snadv has three main goals}{

 - Check the model fit by visualizing pearson residuals, qq plots on rescaled time events and intensity 
   plots.
 - Evaluate the model by calculating the log likelihood and residuals.
 - Simulate MMHP/HP process for given parameter.
}

