% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniHawkesPearsonResidual.R
\name{uniHawkesPearsonResidual}
\alias{uniHawkesPearsonResidual}
\title{Return the Pearson residual of the Hawkes Process}
\usage{
uniHawkesPearsonResidual(object, events, termination)
}
\arguments{
\item{object}{parameters for Hawkes process, include lambda0, alpha, beta}

\item{events}{vector of event happening time}

\item{termination}{termination time}
}
\description{
Take a Hawkes Process object including lambda0, alpha and beta, as well as the event times, then generate its pearson residual
}
