% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniMMHPIntensity.R
\name{UniMMHPIntensity}
\alias{UniMMHPIntensity}
\title{Return the intensity function of the Markov-modulated Hawkes Process(MMHP)}
\usage{
UniMMHPIntensity(mmhp, simulation)
}
\arguments{
\item{mmhp}{a mmhp object including its state, state_time, tau, lambda0, lambda1, beta and alpha.}

\item{simulation}{the simulated Markov-modulated Hawkes Process(MMHP)}
}
\description{
Take a mmhp object and generate its intensity function accordingly
}
\examples{
Q <- matrix(c(-0.4, 0.4, 0.2, -0.2), ncol = 2, byrow = TRUE)
x <- mmhp(Q, delta = c(1 / 3, 2 / 3), lambda0 = 0.9, lambda1 = 1.1, alpha = 0.8, beta = 1.2)
y <- simulatemmhp(x)
z <- UniMMHPIntensity(x, y)
}
