% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPIntensity.R
\name{HPIntensity}
\alias{HPIntensity}
\title{Return the intensity function of Hawkes Process}
\usage{
HPIntensity(lambda, i, alpha, beta, start, end, history, hawkes_time)
}
\arguments{
\item{lambda}{parameters for Hawkes process}

\item{i}{state number}

\item{alpha}{parameters for Hawkes process}

\item{beta}{parameters for Hawkes process}

\item{start}{the start time of current state}

\item{end}{the end time of current state}

\item{history}{the past event times}

\item{hawkes_time}{the event times happened in this state}
}
\description{
Return the intensity function of Hawkes Process, a helper function for 'UniMMHPIntensity'
}
