% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snadv-package.R
\docType{package}
\name{snadv}
\alias{snadv}
\alias{snadv-package}
\title{snadv: Perform diagnosis and visualizations for social network analysis}
\description{
The model checking process is an important part in the social network analysis.
This package includes multiple methods to plot the model as well as the true data, to compare the model fit and model performace.
}
